bot/requirements.txt
claude-agent-sdk>=0.1.0
python-telegram-bot>=21.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
